These changes are in order to add hdfs csv support to the datasource.
Changed datasource SQL generation to add casting
CAST(COL AS TYPE), along with options to
 1) disable casting "DisableCasts"
 2) disable certain casts "DisabledCasts", a comma separated list of
    types to disable cast for.  Example: "TIMESTAMP,NUMERIC"
Casting is done using the schema dataType and the type encoded
in the SQL Query adheres to AWS S3 Select spec.

Also made changes around removing quotes where not needed, including
around column names (select or filter), in Group By column names.

Added new Pushdown object routine getReadableQuery(), which
allows for returning a query string that is more readable by
eliminating casts and by using column names.

Also removed info tracing from Pushdown.queryFromSchema in favor of
letting the clients do it as desired.
Added tracing in hdfs and s3 clients.

Also added ability for hdfs data source to skip header rows.
This was needed in the case where we are not pushing down,
and choose to use hdfs directly.  In this case we need
to skip the first row in the first partition.